Smooth Transition bugs fixed with unchangable MIN_INTER_CAR_SPACE hack TODO: Random Turns after each road. Scaling of Minimum inter-car distance.

+++ src/car.h
@@ -17,7 +17,7 @@ class car
 {
  public:
   car (road* init_road, int next_turn);
-  car (road* init_road, int next_turn, int _position);
+  car (road* init_road, int next_turn, float _position);
   void  setCar (road* init_road, int next_turn, float _position);
   void write_state(FILE* output);
   bool can_move();
